+++ CHANGELOG.md
@@ -28,7 +28,12 @@ or major refactoring improvements.
 ### New
 
 --------------------------------------------------------------------------------
-## [0.4.4] - 2024-xx-xx - Enhance Docker
+## [0.4.5] - 2024-02-14 - Show Sutime Pipeline
+### New
+- Produce json extraction at pipeline steps
+- Display text/html with highlight from json
+
+## [0.4.4] - 2024-01-15 - Enhance Docker
 ### Changed
 - Simplified Docker container build
 ### New